refactor: Extract verification system to a module folder
- Created a "bot_modules" system where events, interactions, commands,
etc are organized by module.
- Extracted everything related to the verification system to a folder
under bot_modules.
- Edited the handler.js file to be compatible with this change.

The idea is to modularize the rest of the code base in future commits.

Issue: None
